Moved switch classifiers to file-based format
