+++ bci_essentials/classifiers/__init__.py
@@ -10,3 +10,5 @@ from .mi_classifier import MI_classifier
 from .ssvep_riemmanian_mdm_classifier import SSVEP_riemannian_mdm_classifier
 from .ssvep_basic_tf_classifier import SSVEP_basic_tf_classifier
 from .null_classifier import Null_classifier
+from .switch_mdm_classifier import Switch_mdm_classifier
+from .switch_deep_classifier import Switch_deep_classifier
